refactor App.jsx code
